AssetManager integrated, asset loading refactored
Note this is via SchuAssetManager, a custom implementation that allows for fuzzy searching and aliases.

Fontbook and Audiobox are now relegated to single items part of the SAM to avoid loading redundancy. This took a long time to get right, due to trying not to have to pass around an add'l parameter all the time. Tough.

Lastly shifted asset loading to during LoadScreen, avoiding the initial slow boot (saves ~15s).