Intervals and BatchGroup integrated, phase bugfixes, Fontbook Orientation/cover awareness update
Introduced the Interval system, a replacement for null-checking and boolean-spamming w/ regular Instants — this allows for tighter control, a toggleable flag, and "just" checkers.

The IntervalBuffer should be used for querying actions for later dependent on an interval ending — it is used so far with the client to delay next phase until the interphase buffer ends.

 Various new utilities have been added, such as EasingFunction including smooth(er)step and cubic bezier for greater control. Also a heckuva lot of Interlerper suite bugfixes (a LOT). Orientations! These are now used by Fontbook to be able to context-aware draw text below pre-existing text and anchor them to corners, so no more match-name label shenanigans. And lovely lil' smilies! (^▿^)

Since a certain stupid lil programmer didn't mention it (looking over here...), BatchGroup has been added and provides greater control over rendering and manipulating Batchables. This version adds CSS-like z-index, so oughta be a good place to mention it eh?