Focusables ported, 3D debug mesh tool added, interlerper improved
Ported Focusable and FocusableGroup from Gambler's Delight, with a few improvements. This version acts similarly to step(), which is different from its predecessor which could rely on just one-and-done Transition. Some keyboard controls (including Focusable) has been added to menu and ready prompt as a test.

A new debug tool for 3D has been added; you can use this to (a) test camera paths and (b) create real-time meshes. And hey, it supports camera interpolation!! This is a good practice/basis for future additions to the game!

Interlerper now can be instantiated by a certain step count (effectively "scaled time"), which was used for debug cam interp but will be quite useful later.