Wiimote support added, EvictingList/EvictingMap and debug render optimisations introduced, new assets
Ok... technically not quite, but this is the most beautiful commit I've ever seen :)

gdx-controller added, tested with Joycons but couldn't test Wiimote due to Ventura shenanigans. It *theoretically* works!

EvictingList and EvictingMap added, which are effectively circular FIFO queues but a huge dependency isn't needed to use it like, once...

...and they are used for some incredible debug render optimisations — decals and models now limit cached, so no more 5+ frame buffer rendering per frame! That's 20 -> 120 FPS!